Fixing json key error

+++ modules/main/sockets.py
@@ -20,32 +20,34 @@ def disconnect():
 
 @socketio.on('mapHashID')
 def mapHashID(Hash):
-    pika_client = pika.BlockingConnection(pika.URLParameters(MQ_URL))
-    channel = pika_client.channel()
 
-    if redis_client.exists(Hash):
-        redis_client.delete(redis_client.get(Hash))
-    redis_client.set(request.sid,Hash)
-    redis_client.set(Hash,request.sid)
+    if Hash != None:
+        pika_client = pika.BlockingConnection(pika.URLParameters(MQ_URL))
+        channel = pika_client.channel()
 
-    queue_val = hash_func(Hash)
-    all_msgs = []
-    val = channel.queue_declare(queue=str(queue_val),passive=True)
-    num_msgs = val.method.message_count
-    if num_msgs != 0:
-        for method_frame,_,body in channel.consume(str(queue_val)):
-            body = body.decode('utf-8')
-            user_msg = json.loads(body)
-            if user_msg['friendHashID'] == Hash:
-                all_msgs.append(user_msg)
-                channel.basic_ack(method_frame.delivery_tag)
-            num_msgs = num_msgs - 1
+        if redis_client.exists(Hash):
+            redis_client.delete(redis_client.get(Hash))
+        redis_client.set(request.sid,Hash)
+        redis_client.set(Hash,request.sid)
 
-            if num_msgs == 0:
-                break
-        all_msgs = json.dumps(all_msgs)
-        emit('unread',all_msgs)
-        print(all_msgs)
-    channel.close()
+        queue_val = hash_func(Hash)
+        all_msgs = []
+        val = channel.queue_declare(queue=str(queue_val),passive=True)
+        num_msgs = val.method.message_count
+        if num_msgs != 0:
+            for method_frame,_,body in channel.consume(str(queue_val)):
+                body = body.decode('utf-8')
+                user_msg = json.loads(body)
+                if user_msg['friendHashID'] == Hash:
+                    all_msgs.append(user_msg)
+                    channel.basic_ack(method_frame.delivery_tag)
+                num_msgs = num_msgs - 1
+
+                if num_msgs == 0:
+                    break
+            all_msgs = json.dumps(all_msgs)
+            emit('unread',all_msgs)
+            print(all_msgs)
+        channel.close()
